Poloshing inspection import screens
